+++ tools/iceworks-cli/bin/iceworks.js
@@ -25,29 +25,23 @@ program.arguments('<command>').action((cmd) => {
 });
 
 program
-  .command('init')
+  .command('init [npmName]')
   .description('init project by template')
-  .option(
-    '-t, --template <template>',
-    'Specify the npm package name for the template'
-  )
   .on('--help', () => {
     console.log('');
     console.log('Examples:');
     console.log('  $ iceworks init');
-    console.log('  $ iceworks init -t @icedesign/lite-scaffold');
+    console.log('  $ iceworks init @icedesign/lite-scaffold');
   })
-  .action((cmd) => {
-    require('../command/init')(cleanArgs(cmd));
+  .action((npmName, cmd) => {
+    const options = cleanArgs(cmd);
+    options.npmName = npmName;
+    require('../command/init')(options);
   });
 
 program
-  .command('add [type]')
+  .command('add <npmName>')
   .description('add block to current directory')
-  .option(
-    '-t, --template <template>',
-    'Specify the npm package name of the block'
-  )
   .option(
     '-n, --name <name>',
     'Specify the block directory name like CustomBlock'
@@ -55,12 +49,13 @@ program
   .on('--help', () => {
     console.log('');
     console.log('Examples:');
-    console.log('  $ iceworks add block');
-    console.log('  $ iceworks add block -t @icedesign/user-landing-block');
-    console.log('  $ iceworks add block -t @icedesign/user-landing-block -n CustomBlock');
+    console.log('  $ iceworks add @icedesign/user-landing-block');
+    console.log('  $ iceworks add @icedesign/user-landing-block -n CustomBlock');
   })
-  .action((type, cmd) => {
-    require('../command/addBlock')(cleanArgs(cmd));
+  .action((npmName, cmd) => {
+    const options = cleanArgs(cmd);
+    options.npmName = npmName;
+    require('../command/addBlock')(options);
   });
 
 // add some useful info on help